+++ boss.py
@@ -1,0 +1,4 @@
+from enemy import Enemy
+
+class boss(Enemy):
+
